modify middleware.go
