+++ pkg/middleware/quota.go
@@ -2,7 +2,7 @@ package middleware
 
 import "github.com/gin-gonic/gin"
 
-func registerIpAllowListHandler() gin.HandlerFunc {
+func registerQuotaHandler() gin.HandlerFunc {
 	return func(c *gin.Context) {
 
 	}
